+++ src/main/java/ladder/domain/Position.java
@@ -1,4 +1,4 @@
-package ladder.domain.participant;
+package ladder.domain;
 
 public record Position(int value) {
 }
